+++ src/main/java/net/sf/mzmine/modules/rawdatamethods/extractscans/ExtractScansParameters.java
@@ -16,7 +16,7 @@
  * USA
  */
 
-package net.sf.mzmine.modules.rawdatamethods.exportscans;
+package net.sf.mzmine.modules.rawdatamethods.extractscans;
 
 import java.text.DecimalFormat;
 import net.sf.mzmine.parameters.Parameter;
